:truck: [open-formulieren/open-forms#4929] Move SubmissionProvider into its own component
The component was moved from Form.jsx into its own
file and imports are updated.
